Commit modified version of our private website
This is a modified version of an existing website I previously made for our private Discord server. It is made to be public, so some details have been anonymized, such as our group name, YouTube channel and statistics. The data will be fetched from another source, but works just like it does on the private website.